Extract CopyToClipboard to own component
