+++ app/pages/project/instances/instance/tabs/NetworkingTab.tsx
@@ -27,6 +27,7 @@ import {
   Success12Icon,
 } from '@oxide/ui'
 
+import { CopyToClipboard } from 'app/components/CopyToClipboard'
 import CreateNetworkInterfaceForm from 'app/forms/network-interface-create'
 import EditNetworkInterfaceForm from 'app/forms/network-interface-edit'
 import {
@@ -84,7 +85,18 @@ function ExternalIpsFromInstanceName({ value: primary }: { value: boolean }) {
     query: { project },
   })
   const ips = data?.items.map((eip) => eip.ip).join(', ')
-  return <span className="text-secondary">{primary ? ips : <>&mdash;</>}</span>
+  return (
+    <div className="flex w-full items-center justify-between text-secondary">
+      {primary ? (
+        <>
+          <span>{ips}</span>
+          <CopyToClipboard text={`${ips}`} />
+        </>
+      ) : (
+        <>&mdash;</>
+      )}
+    </div>
+  )
 }
 
 NetworkingTab.loader = async ({ params }: LoaderFunctionArgs) => {
